+++ src/commands.ts
@@ -1,6 +1,8 @@
 import { $, ShellPromise } from "bun";
 import { PullRequestFactory } from "./models/PullRequestFactory";
-import { PullRequestResponse } from "./models/GitHubResponse";
+import { PullRequestResponse, UserResponse } from "./models/GitHubResponse";
+import { User } from "./models/User";
+import { UserFactory } from "./models/UserFactory";
 
 const makeGhJsonRequest = async <T>(
   command: ShellPromise,
@@ -38,7 +40,9 @@ export const prFields = [
   "mergeStateStatus",
 ] as const;
 
-export const fetchMentionedPrs = async (repo: string) => {
+export const userFields = ["login", "id"] as const;
+
+export const fetchMentionedPrs = async (repo: string, activeUser: User) => {
   const command = $`gh pr list --search "mentions:@me -author:@me" --repo $REPO --json $FIELDS`;
 
   const pullRequests = await makeGhJsonRequest<PullRequestResponse[]>(command, {
@@ -46,10 +50,10 @@ export const fetchMentionedPrs = async (repo: string) => {
     FIELDS: prFields.join(","),
   });
 
-  return pullRequests.map(PullRequestFactory.from);
+  return pullRequests.map((pr) => PullRequestFactory.from(pr, activeUser));
 };
 
-export const fetchReviewedPrs = async (repo: string) => {
+export const fetchReviewedPrs = async (repo: string, activeUser: User) => {
   const command = $`gh pr list --search "reviewed-by:@me -author:@me" --repo $REPO --json $FIELDS`;
 
   const pullRequests = await makeGhJsonRequest<PullRequestResponse[]>(command, {
@@ -57,10 +61,10 @@ export const fetchReviewedPrs = async (repo: string) => {
     FIELDS: prFields.join(","),
   });
 
-  return pullRequests.map(PullRequestFactory.from);
+  return pullRequests.map((pr) => PullRequestFactory.from(pr, activeUser));
 };
 
-export const fetchMyPullRequests = async (repo: string) => {
+export const fetchMyPullRequests = async (repo: string, activeUser: User) => {
   const command = $`gh pr list --repo $REPO --author="@me" --json $FIELDS`;
 
   const pullRequests = await makeGhJsonRequest<PullRequestResponse[]>(command, {
@@ -68,12 +72,14 @@ export const fetchMyPullRequests = async (repo: string) => {
     FIELDS: prFields.join(","),
   });
 
-  return pullRequests.map(PullRequestFactory.from);
+  return pullRequests.map((pr) => PullRequestFactory.from(pr, activeUser));
 };
 
 export const fetchRequestingReviewPullRequests = async (
   repo: string,
   labels: string[],
+  teamMembers: User[],
+  activeUser: User,
 ) => {
   const command = $`gh pr list --repo $REPO --search $SEARCH --json $FIELDS`;
   let queries = ["review-requested:@me"];
@@ -82,13 +88,25 @@ export const fetchRequestingReviewPullRequests = async (
     queries.push(`label:${labels.join(",")}`);
   }
 
+  teamMembers.forEach((member) => {
+    if (labels.length > 0) {
+      queries.push(
+        `OR -reviewed-by:@me -author:@me label:${labels.join(",")} reviewed-by:${member.login}`,
+      );
+    } else {
+      queries.push(
+        `OR -reviewed-by:@me -author:@me reviewed-by:${member.login}`,
+      );
+    }
+  });
+
   const pullRequests = await makeGhJsonRequest<PullRequestResponse[]>(command, {
     REPO: repo,
     FIELDS: prFields.join(","),
     SEARCH: `${queries.join(" ")}`,
   });
 
-  return pullRequests.map(PullRequestFactory.from);
+  return pullRequests.map((pr) => PullRequestFactory.from(pr, activeUser));
 };
 
 export const fetchLatestRelease = async (): Promise<string | undefined> => {
@@ -103,4 +121,24 @@ export const fetchLatestRelease = async (): Promise<string | undefined> => {
   >(command, { FIELDS: fields.join(",") });
 
   return releases.find((release) => !release.isPrerelease)?.tagName;
+};
+
+export const fetchTeamMembers = async (
+  org: string,
+  team: string,
+): Promise<User[]> => {
+  const command = $`gh api orgs/$ORG/teams/$TEAM/members`;
+  const response = await makeGhJsonRequest<UserResponse[]>(command, {
+    ORG: org,
+    TEAM: team,
+  });
+
+  return response.map(UserFactory.from);
+};
+
+export const getActiveGithubUser = async (): Promise<User> => {
+  const command = $`gh api user`;
+  const user = await makeGhJsonRequest<UserResponse>(command, {});
+
+  return UserFactory.from(user);
 };