+++ src/models/GitHub.ts
@@ -67,3 +67,42 @@ export interface GitHubPullRequest {
   mergeStateStatus: "BLOCKED" | "CLEAN" | "DIRTY";
 }
 
+export interface GitHubUser {
+  avatar_url: string;
+  events_url: string;
+  followers_url: string;
+  following_url: string;
+  gists_url: string;
+  gravatar_id: string;
+  html_url: string;
+  id: number;
+  login: string;
+  node_id: string;
+  organizations_url: string;
+  received_events_url: string;
+  repos_url: string;
+  site_admin: boolean;
+  starred_url: string;
+  subscriptions_url: string;
+  type: string;
+  url: string;
+  user_view_type: string;
+}
+
+export interface GitHubMyUser extends GitHubUser {
+  name: string;
+  company: string | null;
+  blog: string;
+  location: string | null;
+  email: string;
+  hireable: boolean | null;
+  bio: string | null;
+  twitter_username: string | null;
+  notification_email: string;
+  public_repos: number;
+  public_gists: number;
+  followers: number;
+  following: number;
+  created_at: string;
+  updated_at: string;
+}