+++ src/models/GitHubResponse.ts
@@ -1,8 +1,11 @@
-import { prFields } from "../commands";
-import { GitHubPullRequest } from "./GitHub";
+import { prFields, userFields } from "../commands";
+import { GitHubPullRequest, GitHubUser } from "./GitHub";
 
 type PullRequestStatusResponseFields = (typeof prFields)[number];
 export type PullRequestResponse = Pick<
   GitHubPullRequest,
   PullRequestStatusResponseFields
 >;
+
+type UserResponseFields = (typeof userFields)[number];
+export type UserResponse = Pick<GitHubUser, UserResponseFields>;